HappyMolecules.LennardJones — Type

\[f(r) = \frac{48 \vec{r}}{r^2}(\frac{1}{r^{12}} - 0.5 \frac{1}{r^6})\]

HappyMolecules.heat_capacity — Method

Compute the constant volume capacity is using the following equation

\[\langle K^2 \rangle - \langle K \rangle^2 = \frac{3 k_b^2 T^2}{2N}(1-\frac{3k_B}{2C_v})\]

HappyMolecules.mean_potential_energy — Method

mean_potential_energy(
    md::HappyMolecules.MDRuntime{D, T}
) -> Any

HappyMolecules.pressure — Method

The most common among the ways to measure the pressure $P$ is based on the virial equation for the pressure.

\[P = \rho k_B T + \frac{1}{dV}\langle\sum_{i<j} f(r_{ij}) \cdot r_{ij}\rangle\]

HappyMolecules.temperature — Method

The temperature $T$ is measured by computing the average kinetic energy per degree of freedom.

\[k_B T = \frac{\langle 2 \mathcal{K} \rangle}{f}.\]